# Parser: inline unnecessary function calls